Create and move page components
