+++ src/component/Header/Header.js
@@ -1,7 +1,7 @@
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faMagnifyingGlass, faBars } from "@fortawesome/free-solid-svg-icons";
 import { useNavigate } from "react-router";
-import * as S from "../pages/MyPage/style";
+import * as S from "./style";
 import "../pages/MyPage/index.css";
 
 const Header = ({ menu }) => {
