+++ src/component/Header/MypageItem.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import * as S from "../pages/MyPage/style";
+import * as S from "./style";
 import { HeartOutlined, HeartFilled } from "@ant-design/icons";
 import { render } from "@testing-library/react";
