+++ src/pages/404ErrorPage/ErrorPage.jsx
@@ -1,6 +1,6 @@
 import * as S from "./style";
 import React from "react";
-import Header from "../../component/Header";
+import Header from "../../component/Header/Header";
 import "./index.css";
 
 function ErrorPage() {
